SFUwiki/requirements.txt
asgiref==3.7.2
coverage==7.3.2
Django==4.2.5
djangorestframework==3.14.0
drf-nested-routers==0.93.4
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
